+++ src/cluster-tile-index.ts
@@ -1,86 +1,20 @@
 import KDBush from 'kdbush';
 import {projectX, projectY} from './convert';
-import {unprojectX, unprojectY} from './deconvert';
+import {unprojectX, unprojectY, featureToGeoJSON} from './deconvert';
+import type {ClusterFeature, ClusterOrPointFeature, ClusterProperties, GeoJSONVTTileIndex, GeoJSONVTFeature, GeoJSONVTInternalFeature, GeoJSONVTInternalPointFeature, GeoJSONVTOptions, GeoJSONVTTile, SuperclusterOptions} from './definitions';
 
-export type SuperclusterOptions = {
-    /**
-     * Min zoom to generate clusters on
-     * @default 0
-     */
-    minZoom?: number;
-    /**
-     * Max zoom level to cluster the points on
-     * @default 16
-     */
-    maxZoom?: number;
-    /**
-     * Minimum points to form a cluster
-     * @default 2
-     */
-    minPoints?: number;
-    /**
-     * Cluster radius in pixels
-     * @default 40
-     */
-    radius?: number;
-    /**
-     * Tile extent (radius is calculated relative to it)
-     * @default 512
-     */
-    extent?: number;
-    /**
-     * Size of the KD-tree leaf node, affects performance
-     * @default 64
-     */
-    nodeSize?: number;
-    /**
-     * Whether to log timing info
-     * @default false
-     */
-    log?: boolean;
-    /**
-     * Whether to generate numeric ids for input features (in vector tiles)
-     * @default false
-     */
-    generateId?: boolean;
-    /**
-     * A reduce function for calculating custom cluster properties
-     * @default null
-     */
-    reduce?: ((accumulated: Record<string, unknown>, props: Record<string, unknown>) => void) | null;
-    /**
-     * Properties to use for individual points when running the reducer
-     * @default props => props
-     */
-    map?: (props: GeoJSON.GeoJsonProperties) => Record<string, unknown>;
+type ClusterFeatureInternal = GeoJSONVTInternalPointFeature & {
+    tags: ClusterProperties;
 };
 
-export type ClusterProperties = {
-    cluster: true;
-    cluster_id: number;
-    point_count: number;
-    point_count_abbreviated: string | number;
-    [key: string]: unknown;
-};
+type ClusterOrPointFeatureInternal = ClusterFeatureInternal | GeoJSONVTInternalPointFeature;
 
-export type ClusterFeature = GeoJSON.Feature<GeoJSON.Point, ClusterProperties>;
-
-export type SuperclusterTileFeature = {
-    type: 1;
-    geometry: [[number, number]];
-    tags: GeoJSON.GeoJsonProperties | ClusterProperties;
-    id?: number | string;
-};
-
-export type SuperclusterTile = {
-    features: SuperclusterTileFeature[];
-};
-
+/** @internal */
 export type KDBushWithData = KDBush & {
     flatData: number[];
 };
 
-const defaultOptions: Required<SuperclusterOptions> = {
+export const defaultClusterOptions: Required<SuperclusterOptions> = {
     minZoom: 0,
     maxZoom: 16,
     minPoints: 2,
@@ -99,15 +33,18 @@ const OFFSET_PARENT = 4;
 const OFFSET_NUM = 5;
 const OFFSET_PROP = 6;
 
-export class Supercluster {
+/**
+ * This class allow clustering of geojson points.
+ */
+export class ClusterTileIndex implements GeoJSONVTTileIndex {
     options: Required<SuperclusterOptions>;
     trees: KDBushWithData[];
     stride: number;
     clusterProps: Record<string, unknown>[];
-    points: GeoJSON.Feature<GeoJSON.Point>[];
+    points: GeoJSONVTInternalPointFeature[];
 
     constructor(options?: SuperclusterOptions) {
-        this.options = Object.assign(Object.create(defaultOptions), options) as Required<SuperclusterOptions>;
+        this.options = Object.assign(Object.create(defaultClusterOptions), options) as Required<SuperclusterOptions>;
         this.trees = new Array(this.options.maxZoom + 1);
         this.stride = this.options.reduce ? 7 : 6;
         this.clusterProps = [];
@@ -115,10 +52,61 @@ export class Supercluster {
     }
 
     /**
-     * Loads input point features and builds the internal clustering index.
-     * @param points - Input GeoJSON point features to cluster.
+     * Loads GeoJSON point features and builds the internal clustering index.
+     * @param points - GeoJSON point features to cluster.
      */
     load(points: GeoJSON.Feature<GeoJSON.Point>[]): void {
+        const features: GeoJSONVTInternalPointFeature[] = [];
+        
+        // Convert GeoJSON point features to GeoJSONVT internal point features
+        for (const point of points) {
+            if (!point.geometry) {
+                continue;
+            }
+
+            const [lng, lat] = point.geometry.coordinates;
+            const [x, y] = [projectX(lng), projectY(lat)];
+            
+            const feature: GeoJSONVTInternalPointFeature = {
+                id: point.id,
+                type: 'Point',
+                geometry: [x, y],
+                tags: point.properties
+            };
+            
+            features.push(feature);
+        }
+        
+        this.createIndex(features);
+    }
+
+    /**
+     * @internal
+     * Loads internal GeoJSONVT point features from a data source and builds the clustering index.
+     * @param features - {@link GeoJSONVTInternalFeature} data source features to filter and cluster.
+     */
+    initialize(features: GeoJSONVTInternalFeature[]): void {
+        const points: GeoJSONVTInternalPointFeature[] = [];
+
+        for (const feature of features) {
+            if (feature.type !== 'Point') continue;
+            points.push(feature);
+        }
+
+        this.createIndex(points);
+    }
+
+    /**
+     * @internal
+     * Updates the cluster data by rebuilding.
+     * @param features 
+     */
+    updateIndex(features: GeoJSONVTInternalFeature[], _affected: GeoJSONVTInternalFeature[], options: GeoJSONVTOptions) {
+        this.options = Object.assign(Object.create(defaultClusterOptions), options.clusterOptions) as Required<SuperclusterOptions>;
+        this.initialize(features);
+    }
+
+    private createIndex(points: GeoJSONVTInternalPointFeature[]): void {
         const {log, minZoom, maxZoom} = this.options;
 
         if (log) console.time('total time');
@@ -133,11 +121,12 @@ export class Supercluster {
 
         for (let i = 0; i < points.length; i++) {
             const p = points[i];
-            if (!p.geometry) continue;
+            if (!p?.geometry) continue;
 
-            const [lng, lat] = p.geometry.coordinates;
-            const x = Math.fround(projectX(lng));
-            const y = Math.fround(projectY(lat));
+            let [x, y] = p.geometry;
+            x = Math.fround(x);
+            y = Math.fround(y);
+
             // store internal point/cluster data in flat numeric arrays for performance
             data.push(
                 x, y, // projected point coordinates
@@ -171,7 +160,12 @@ export class Supercluster {
      * @param bbox - Bounding box in `[westLng, southLat, eastLng, northLat]` order.
      * @param zoom - Zoom level to query.
      */
-    getClusters(bbox: [number, number, number, number], zoom: number): (ClusterFeature | GeoJSON.Feature<GeoJSON.Point>)[] {
+    public getClusters(bbox: [number, number, number, number], zoom: number): ClusterOrPointFeature[] {
+        const clusterInternal = this.getClustersInternal(bbox, zoom);
+        return clusterInternal.map((f) => featureToGeoJSON(f) as ClusterOrPointFeature);
+    }
+
+    private getClustersInternal(bbox: [number, number, number, number], zoom: number): ClusterOrPointFeatureInternal[] {
         let minLng = ((bbox[0] + 180) % 360 + 360) % 360 - 180;
         const minLat = Math.max(-90, Math.min(90, bbox[1]));
         let maxLng = bbox[2] === 180 ? 180 : ((bbox[2] + 180) % 360 + 360) % 360 - 180;
@@ -181,27 +175,27 @@ export class Supercluster {
             minLng = -180;
             maxLng = 180;
         } else if (minLng > maxLng) {
-            const easternHem = this.getClusters([minLng, minLat, 180, maxLat], zoom);
-            const westernHem = this.getClusters([-180, minLat, maxLng, maxLat], zoom);
+            const easternHem = this.getClustersInternal([minLng, minLat, 180, maxLat], zoom);
+            const westernHem = this.getClustersInternal([-180, minLat, maxLng, maxLat], zoom);
             return easternHem.concat(westernHem);
         }
 
         const tree = this.trees[this.limitZoom(zoom)];
         const ids = tree.range(projectX(minLng), projectY(maxLat), projectX(maxLng), projectY(minLat));
         const data = tree.flatData;
-        const clusters: (ClusterFeature | GeoJSON.Feature<GeoJSON.Point>)[] = [];
+        const clusters: ClusterOrPointFeatureInternal[] = [];
         for (const id of ids) {
             const k = this.stride * id;
-            clusters.push(data[k + OFFSET_NUM] > 1 ? getClusterJSON(data, k, this.clusterProps) : this.points[data[k + OFFSET_ID]]);
+            clusters.push(data[k + OFFSET_NUM] > 1 ? getClusterFeature(data, k, this.clusterProps) : this.points[data[k + OFFSET_ID]]);
         }
         return clusters;
     }
 
     /**
-     * Returns the immediate children (clusters or points) of a cluster.
+     * Returns the immediate children (clusters or points) of a cluster as GeoJSON.
      * @param clusterId - The target cluster id.
      */
-    getChildren(clusterId: number): (ClusterFeature | GeoJSON.Feature<GeoJSON.Point>)[] {
+    getChildren(clusterId: number): ClusterOrPointFeature[] {
         const originId = this.getOriginId(clusterId);
         const originZoom = this.getOriginZoom(clusterId);
         const clusterError = new Error('No cluster with the specified id: ' + clusterId);
@@ -216,11 +210,11 @@ export class Supercluster {
         const x = data[originId * this.stride];
         const y = data[originId * this.stride + 1];
         const ids = tree.within(x, y, r);
-        const children: (ClusterFeature | GeoJSON.Feature<GeoJSON.Point>)[] = [];
+        const children: ClusterOrPointFeature[] = [];
         for (const id of ids) {
             const k = id * this.stride;
             if (data[k + OFFSET_PARENT] === clusterId) {
-                children.push(data[k + OFFSET_NUM] > 1 ? getClusterJSON(data, k, this.clusterProps) : this.points[data[k + OFFSET_ID]]);
+                children.push(data[k + OFFSET_NUM] > 1 ? getClusterGeoJSON(data, k, this.clusterProps) : featureToGeoJSON(this.points[data[k + OFFSET_ID]]) as GeoJSON.Feature<GeoJSON.Point>);
             }
         }
 
@@ -251,7 +245,7 @@ export class Supercluster {
      * @param x - Tile x coordinate.
      * @param y - Tile y coordinate.
      */
-    getTile(z: number, x: number, y: number): SuperclusterTile | null {
+    getTile(z: number, x: number, y: number): GeoJSONVTTile | null {
         const tree = this.trees[this.limitZoom(z)];
         const z2 = Math.pow(2, z);
         const {extent, radius} = this.options;
@@ -259,8 +253,13 @@ export class Supercluster {
         const top = (y - p) / z2;
         const bottom = (y + 1 + p) / z2;
 
-        const tile: SuperclusterTile = {
-            features: []
+        const tile: GeoJSONVTTile = {
+            transformed: true,
+            features: [],
+            source: null,
+            x: x,
+            y: y,
+            z: z
         };
 
         this.addTileFeatures(
@@ -326,7 +325,7 @@ export class Supercluster {
         return tree;
     }
 
-    private addTileFeatures(ids: number[], data: number[], x: number, y: number, z2: number, tile: SuperclusterTile): void {
+    private addTileFeatures(ids: number[], data: number[], x: number, y: number, z2: number, tile: GeoJSONVTTile): void {
         for (const i of ids) {
             const k = i * this.stride;
             const isCluster = data[k + OFFSET_NUM] > 1;
@@ -340,13 +339,11 @@ export class Supercluster {
                 py = data[k + 1];
             } else {
                 const p = this.points[data[k + OFFSET_ID]];
-                tags = p.properties;
-                const [lng, lat] = p.geometry.coordinates;
-                px = projectX(lng);
-                py = projectY(lat);
+                tags = p.tags;
+                [px, py] = p.geometry;
             }
 
-            const f: SuperclusterTileFeature = {
+            const f: GeoJSONVTFeature = {
                 type: 1,
                 geometry: [[
                     Math.round(this.options.extent * (px * z2 - x)),
@@ -472,13 +469,22 @@ export class Supercluster {
             const props = this.clusterProps[data[i + OFFSET_PROP]];
             return clone ? Object.assign({}, props) : props;
         }
-        const original = this.points[data[i + OFFSET_ID]].properties;
+        const original = this.points[data[i + OFFSET_ID]].tags;
         const result = this.options.map(original);
         return clone && result === original ? Object.assign({}, result) : result;
     }
 }
 
-function getClusterJSON(data: number[], i: number, clusterProps: Record<string, unknown>[]): ClusterFeature {
+function getClusterFeature(data: number[], i: number, clusterProps: Record<string, unknown>[]): ClusterFeatureInternal {
+    return {
+        id: data[i + OFFSET_ID],
+        type: 'Point',
+        tags: getClusterProperties(data, i, clusterProps),
+        geometry: [data[i], data[i + 1]]
+    };
+}
+
+function getClusterGeoJSON(data: number[], i: number, clusterProps: Record<string, unknown>[]): ClusterFeature {
     return {
         type: 'Feature',
         id: data[i + OFFSET_ID],